refactor: corrigir caminhos de importação para componentes de refeições
